run pint

+++ src/Decorator/InvoicePrice/Invoice.php
@@ -4,7 +4,7 @@ namespace Src\Decorator\InvoicePrice;
 
 class Invoice implements InvoicePrice
 {
-    public function price(): int
+    public function price(): float
     {
         return 100000;
     }
