+++ src/Decorator/InvoicePrice/InvoicePrice.php
@@ -4,5 +4,5 @@ namespace Src\Decorator\InvoicePrice;
 
 interface InvoicePrice
 {
-    public function price(): int;
+    public function price(): float;
 }
