+++ src/Decorator/InvoicePrice/InvoicePriceDecorator.php
@@ -4,11 +4,9 @@ namespace Src\Decorator\InvoicePrice;
 
 class InvoicePriceDecorator implements InvoicePrice
 {
-    public function __construct(private readonly InvoicePrice $invoicePrice)
-    {
-    }
+    public function __construct(private readonly InvoicePrice $invoicePrice) {}
 
-    public function price(): int
+    public function price(): float
     {
         return $this->invoicePrice->price();
     }
