+++ src/Decorator/InvoicePrice/ServiceInvoicePrice.php
@@ -4,7 +4,7 @@ namespace Src\Decorator\InvoicePrice;
 
 class ServiceInvoicePrice extends InvoicePriceDecorator
 {
-    public function price(): int
+    public function price(): float
     {
         return parent::price() + 3200;
     }
