+++ src/Decorator/InvoicePrice/VATInvoicePrice.php
@@ -4,9 +4,9 @@ namespace Src\Decorator\InvoicePrice;
 
 class VATInvoicePrice extends InvoicePriceDecorator
 {
-    public function price(): int
+    public function price(): float
     {
-//        return parent::price() * 1.09;
+        //        return parent::price() * 1.09;
         return parent::price() + (parent::price() * 0.09);
     }
 }
